move metric observerbilty into dedicated file
